Correct hole sizes and align with KLC
The holes for the contact pins are too small to fit the 1mm pins on the
relay, and smaller than the datasheet recommendation of 1.3mm. Revised
to 1.3 mm. Note that datasheet also recommends 1.3mm for coil pins, but
this is unnecessary as these pins are smaller.

Alignment with KLC: Added ref to fab layer and added 3D model path